requirements.txt
pytest
orjson
httpx==0.27.2
respx
pytest-asyncio
greenlet
gevent
click
markdown
pytest-cov
tenacity
PyNaCl
ruamel.yaml
bashlex
lxml